requirements-qt.txt
PySide6==6.11.0
pytest-qt>=4.4
